+++ AndroidAsync/src/com/koushikdutta/async/http/callback/RequestCallback.java
@@ -5,5 +5,5 @@ import com.koushikdutta.async.http.AsyncHttpResponse;
 
 public interface RequestCallback<T> extends ResultCallback<AsyncHttpResponse, T> {
     public void onConnect(AsyncHttpResponse response);
-    public void onProgress(AsyncHttpResponse response, int downloaded, int total);
+    public void onProgress(AsyncHttpResponse response, long downloaded, long total);
 }
